Reconfigured so it redirects to the sign in page after signing out.
Devise redirects to the home page by default after signing out and
displays a notice about successfully signing out. Since users must be
logged in to access the home page of this site, it was redirecting to the
sign in page and displaying an annoying message about needing to sign in
first.