Update to redirect to the sign in page after registration
This fixes a problem with the flash messages. Be default, devise redirects
to the home page. Since the home page is not accessible without signing
in, it would redirect again to the sign in page and display a message
telling the user to sign up or sign in rather than displaying the message
about the email containing a confirmation link.