Bug Fix: Fixes a bug with registration introduced in 9869346.
Changes made to configure_permitted_parameters caused fields not
explicitly permitted not to be permitted when submitting a form.

Based on documention
(https://github.com/plataformatec/devise#strong-parameters), the changes
from 9869346 should have worked because new parameters should be appended
to the list of previously permitted parameters. Since this is not working,
I am changing it back.